discard multiple values returned by handle-request
In default method for `handler-request'

```
(throw 'handler-done
   (values nil cond (get-backtrace)))
```

is used, which is caught in `process-request' and the second return
value is used as error flag. However if handler-request returns
multiple values normally from user code, this also registers as
error.

For example, a define-easy-handler such as

```
(define-easy-handler (...) ()
    ...
    (gethash key table))
```

seems OK as user code. But the `gethash' returns T as second value
which undesirably leads to error.