Fix shutdown acceptor wake when listening on 0.0.0.0. (#139)
* Fix shutdown acceptor wake when listening on 0.0.0.0.

0.0.0.0/8 (and the IPv6 equivalent, ::) are invalid addresses to use as
the target of a connection, but those are the local addresses of an
acceptor listening on them. In order for WAKE-ACCEPTOR-FOR-SHUTDOWN to
connect to such an acceptor, we need to use an alternate address when
that's the case.

This also adds a test to detect the failure to interrupt
USOCKET:WAIT-FOR-INPUT which causes the hang. This runs in the same test
context as the normal confidence tests (to preserve the summary at the
end of the run), but uses its own acceptor to avoid potential state
issues. Test-failure conditions have been refactored slightly to support
more general tests like this.

* Avoid setting *DEFAULT-PATHNAME-DEFAULTS* to a file pathname during tests.

When *D-P-D* is a file pathname, pathnames for files that do not have a
file extension wind up having an extension added by MERGE-PATHNAMES, even
when that is only intended to merge the path component (this may depend on
whether the implementation uses NIL or :UNSPECIFIC as the :TYPE component
when parsing paths that don't have an extension).

In particular, this causes a problem when drakma's user-agent-string code
calls SOFTWARE-VERSION on linux; SBCL's implementation reads the
/proc/sys/kernel/osrelease file, but MERGE-PATHNAMES causes it to try to
open /proc/sys/kernel/osrelease.lisp, which doesn't exist. The test need
the directory component of *D-P-D* set for some file-handling tests, but
they do not rely on the filename portion, so strip that before rebinding
*D-P-D* around the tests.